refactor: update transaction components name
