+++ src/components/TransactionItem/TransactionItemCell/styled.tsx
