+++ src/components/TransactionItem/TransactionItemCellList/styled.tsx
