+++ src/pages/Address/index.tsx
@@ -8,7 +8,7 @@ import OverviewCard, { OverviewItemData } from '../../components/Card/OverviewCa
 import TitleCard from '../../components/Card/TitleCard'
 import Content from '../../components/Content'
 import Tooltip from '../../components/Tooltip'
-import TransactionItem from '../../components/Transaction/TransactionItem/index'
+import TransactionItem from '../../components/TransactionItem/index'
 import { AppContext } from '../../contexts/providers/index'
 import { StateWithDispatch } from '../../contexts/providers/reducer'
 import { getAddress } from '../../service/app/address'
